Split BMC dump manager to accomodate more locally stored dumps
The BMC dump manager is having all infrastucure to manage
locally stored dumps. To avoid code duplication while
implementing other locally stored dumps, splitting the
BMC dump manager into locally stored dump manager and
BMC dump manager.
Changes:
Add a new base class bmc_stored::Manager for bmc::Manager
new base class will have following functions
- Store dump properties
- create dump entry based on file watch.
New class diagram will look like this
          |-------------|
          |  base dump  |
          |   manager   |
          |-------------|
                 |
          |-------------|
          | BMC stored  |
          |   dump      |
          |  manager    |
          |-------------|
                |
     -------------------------
     |                       |
 |---------|             |----------|
 |  BMC    |             |   New    |
 |  dump   |             |  dump    |
 | manager |             | Manager  |
 |---------|             |----------|
